backend/requirements.txt
asgiref==3.2.10
Django==3.0.5
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djongo==1.3.3
dnspython==1.16.0
gunicorn==20.0.4
mongoengine==0.20.0
PyJWT==1.7.1
pymongo==3.10.1
pytz==2020.1
sqlparse==0.2.4
whitenoise==5.1.0
